requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
langchain==0.1.20
langchain-community==0.0.38
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0
faiss-cpu==1.8.0
langchain-anthropic==0.1.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
reportlab==4.1.0
markdown>=3.6
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
tqdm==4.66.1
